Re-organização do código
