+++ sgab-core/src/sgab/model/dto/util/BibliotecaHelper.java
@@ -15,7 +15,7 @@ public class BibliotecaHelper {
             resultado = false;
         }
         
-        else if(biblioteca.getId() <= 0){
+        else if(biblioteca.getId().equals("0")){
             resultado = false;
         }
