+++ sgab-view-web/web/core/biblioteca/gerenciar.jsp
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <title>Gerenciamento Biblioteca</title>
     <meta content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" name="viewport">
-    <link rel="stylesheet" href="template-form.css">
+    <link rel="stylesheet" href="../../css/styles.css">
 </head>
 <body>
     <header>
@@ -28,29 +28,29 @@
             <div id="caixa-form">
                 <div id="alterarInfo" class="divInputs">
                     <h1>Alterar Informações</h1>
-                    <form action="gerenciarServlet.java" method="POST">
+                    <form action="../../../src/java/sgab/controller/biblioteca/gerenciarServlet.java" method="POST">
                         <label for="id">Pesquisar por Id</label>
                         <input type="number" id="alterarNomeInput" name="pesquisarPorId" placeholder="Digite o Id">
                         <button type="submit">Pesquisar</button>
                     </form>
-                    <form action="gerenciarServlet.java" method="POST">
+                    <form action="../../../src/java/sgab/controller/biblioteca/gerenciarServlet.java" method="POST">
                         <label for="id">Pesquisar por Nome</label>
                         <input type="text" id="alterarNomeInput" name="pesquisarPorNome" placeholder="Digite o Nome">
                         <button type="submit">Pesquisar</button>
                     </form>
-                    <form action="gerenciarServlet.java" method="POST">
+                    <form action="../../../src/java/sgab/controller/biblioteca/gerenciarServlet.java" method="POST">
                         <label for="id">Alterar Nome</label>
                         <input type="text" id="alterarNomeInput" name="nome" placeholder="Pesquisa biblioteca pelo Nome">
                         <input type="text" id="alterarNomeInput" name="alterarNomeInput" placeholder="Altera o Nome">
                         <button type="submit">Alterar</button>
                     </form>
-                    <form action="gerenciarServlet.java" method="POST">
+                    <form action="../../../src/java/sgab/controller/biblioteca/gerenciarServlet.java" method="POST">
                         <label for="id">Alterar ID</label>
                         <input type="text" id="alterarNomeInput" name="nome" placeholder="Pesquisa biblioteca pelo Nome">
                         <input type="number" id="alterarIDInput" name="alterarIDInput" placeholder="Altera o ID">
                         <button type="submit">Alterar</button>
                     </form>
-                    <form action="gerenciarServlet.java" method="POST">
+                    <form action="../../../src/java/sgab/controller/biblioteca/gerenciarServlet.java" method="POST">
                         <label for="id">Alterar Org</label>
                         <input type="text" id="alterarNomeInput" name="nome" placeholder="Pesquisa biblioteca pelo Nome">
                         <input type="text" id="alterarOrgInput" name="alterarOrgInput" placeholder="Altera a Organização">
