+++ sgab-view-web/web/core/biblioteca/index.jsp
@@ -7,7 +7,7 @@
         <meta http-equiv="X-UA-Compatible" content="IE=edge">
         <title>SGAB | Sistema de Gestão de Acervo Bibliográfico </title>
         <meta content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" name="viewport">
-        <link rel="stylesheet" href="/sgab/css/styles.css">
+        <link rel="stylesheet" href="../../css/styles.css">
         <script src="/sgab/js/helper.js"></script>
     </head>
     <body>
