--- sgab-view-web/web/opcoes.css
@@ -1,116 +1,0 @@
-body {
-    width: 100%;
-    min-width: 360px;
-    margin: auto;
-    background-color: #FFFFFF;
-    padding: 0;
-    text-align: left;
-}
-
-header {
-    background-color: #004C84;
-    height: 78px;
-    border-bottom: solid 3px #bebebe;
-    display: flex;
-    justify-content: space-between;
-    position: relative;
-}
-
-#pagina-inicial a {
-    text-decoration: none;
-    cursor: pointer;
-    color: #3c8dbc;
-    position: absolute;
-    top: 15px;
-    font-size: 45px;
-    padding: 0px 10px;
-    font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;
-}
-
-#logo-topo {
-    padding-right: 310px;
-}
-
-#logo-topo img {
-    bottom: 0;
-    max-height: 45px;
-    padding: 0px 10px;
-    position: absolute;
-    top: 15px;
-
-}
-
-main {
-    background-color: #F9F9F9;
-}
-
-#nome-form {
-    position: relative;
-    padding: 0;
-    background-color: #E1E1E1;
-}
-
-#nome-form h1 {
-    margin: 0;
-    padding: 10px;
-    font-size: 18px;
-    font-family: 'Patua One', 'Helvetica Neue', 'Arial', sans-serif;
-    line-height: 1.1;
-    color: #337ab7;
-    font-weight: bold;
-}
-
-h1 {
-    margin-right: 100vw;
-}
-
-#form {   
-    min-height: 250px;
-    padding: 15px;
-}
-
-form {
-    margin-right: 2em;
-}
-
-.divInputs {
-    width: 300px;
-}
-
-#caixa-form {
-    display: flex;
-    flex-wrap: wrap;
-    padding: 15px;
-    width: 90%;
-    border-radius: 3px;
-    background: #ffffff;
-    border-top: 3px solid #d2d6de;
-    margin: 0 auto;
-    margin-bottom: 20px;
-    box-shadow: 0 1px 1px rgba(0, 0, 0, 0.1);
-    height: 50%;
-}
-
-button {
-    background-color: #3c8dbc;
-    border-color: #367fa9;
-    border-radius: 8px;
-    box-shadow: none;
-    border: 15px solid transparent;
-    padding: 12px;
-    margin: 10px 0px;
-    color: #fff;
-}
-
-footer {
-    background-color: #004C84;
-    background-image: linear-gradient(to bottom, transparent, rgba(0,0,0,0.3));
-    border-top: solid 3px #bebebe;
-    color: #FFFFFF;
-    display: flex;
-    font-size: 11px;
-    justify-content: space-between;
-    padding-left: 10px;
-    padding-right: 10px;
-    min-height: 20px;
-}
